Move S1MultiBurstSBASNetwork into its own module, update notebook, create S1MultiBurstSBASNetwork tests
